+++ bricks/starter/__brick__/lib/core/env/env.dart
@@ -1,0 +1,11 @@
+import 'package:envied/envied.dart';
+
+part 'env.g.dart';
+
+@Envied(path: '.env')
+abstract final class Env {
+  const Env._();
+
+  @EnviedField(varName: 'API_BASE_URL')
+  static const String apiBaseUrl = _Env.apiBaseUrl;
+}